+++ lib/events/s3sessions/s3handler.go
@@ -451,7 +451,7 @@ func (h *Handler) uploadFile(ctx context.Context, path string, reader io.Reader,
 // is not found.
 func (h *Handler) StreamSessionRecording(ctx context.Context, sessionID session.ID, uploadID string) (io.ReadCloser, error) {
 	if uploadID != "" {
-		return nil, trace.NotImplemented("")
+		return nil, trace.NotImplemented("streaming temporary session recordings not supported for S3")
 	}
 	return h.downloadOriginalFile(ctx, h.recordingPath(sessionID))
 }
